Fixes CC-1922
cherry-pick pull-request #2505
(committed by "zendev port cherry-pick")
